requirements.txt
# Core framework
fastapi==0.116.1
uvicorn==0.35.0
python-multipart
mangum

# Data processing
pandas==2.3.1
numpy==2.3.2
openpyxl==3.1.5

# Document processing
--extra-index-url https://download.pytorch.org/whl/cpu
docling==2.41.0
mammoth==1.9.1
PyPDF2==3.0.1
python_docx==1.1.2
python_pptx==1.0.2
pdf2image==1.17.0

# Image processing (CPU-only OpenCV)
opencv-python-headless==4.10.0.84
Pillow==11.3.0

# Audio processing
pydub==0.25.1

# LangChain ecosystem
langchain==0.3.27
langchain_community==0.3.27
langchain_core==0.3.72
langchain_google_genai==2.1.9
langchain_openai==0.3.28
langsmith==0.4.13

# Google APIs
google_api_python_client==2.173.0
google_auth_oauthlib==1.2.2

# Utilities
loguru==0.7.3
itsdangerous==2.2.0
python-dotenv==1.1.1
python_bcrypt==0.3.2
Requests==2.32.4
setuptools==80.9.0
pydantic==2.11.7
protobuf==6.31.1
decorator

# Streamlit (if needed - consider removing if not used in production)
# streamlit==1.45.1

# Constraints to prevent GPU packages
# --constraint constraints.txt